+++ admin/tool/bulkenrollment/classes/base_form.php
@@ -17,7 +17,7 @@
 /**
  * File containing the base import form.
  *
- * @package    tool_uploadcourse
+ * @package    tool_bulkenrollment
  * @copyright  2013 Frédéric Massart
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
@@ -28,11 +28,11 @@ require_once($CFG->libdir.'/formslib.php');
 /**
  * Base import form.
  *
- * @package    tool_uploadcourse
+ * @package    tool_bulkenrollment
  * @copyright  2013 Frédéric Massart
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
-class tool_uploadcourse_base_form extends moodleform {
+class tool_bulkenrollment_base_form extends moodleform {
 
     /**
      * Empty definition.
@@ -54,45 +54,7 @@ class tool_uploadcourse_base_form extends moodleform {
         $mform->addElement('header', 'importoptionshdr', get_string('importoptions', 'tool_bulkenrollment'));
         $mform->setExpanded('importoptionshdr', true);
 
-        $choices = array(
-            tool_uploadcourse_processor::MODE_CREATE_NEW => get_string('createnew', 'tool_bulkenrollment'),
-            tool_uploadcourse_processor::MODE_CREATE_ALL => get_string('createall', 'tool_bulkenrollment'),
-            tool_uploadcourse_processor::MODE_CREATE_OR_UPDATE => get_string('createorupdate', 'tool_bulkenrollment'),
-            tool_uploadcourse_processor::MODE_UPDATE_ONLY => get_string('updateonly', 'tool_bulkenrollment')
-        );
-        $mform->addElement('select', 'options[mode]', get_string('mode', 'tool_bulkenrollment'), $choices);
-        $mform->addHelpButton('options[mode]', 'mode', 'tool_bulkenrollment');
-
-        $choices = array(
-            tool_uploadcourse_processor::UPDATE_NOTHING => get_string('nochanges', 'tool_bulkenrollment'),
-            tool_uploadcourse_processor::UPDATE_ALL_WITH_DATA_ONLY => get_string('updatewithdataonly', 'tool_bulkenrollment'),
-            tool_uploadcourse_processor::UPDATE_ALL_WITH_DATA_OR_DEFAUTLS =>
-                get_string('updatewithdataordefaults', 'tool_bulkenrollment'),
-            tool_uploadcourse_processor::UPDATE_MISSING_WITH_DATA_OR_DEFAUTLS => get_string('updatemissing', 'tool_bulkenrollment')
-        );
-        $mform->addElement('select', 'options[updatemode]', get_string('updatemode', 'tool_bulkenrollment'), $choices);
-        $mform->setDefault('options[updatemode]', tool_uploadcourse_processor::UPDATE_NOTHING);
-        $mform->hideIf('options[updatemode]', 'options[mode]', 'eq', tool_uploadcourse_processor::MODE_CREATE_NEW);
-        $mform->hideIf('options[updatemode]', 'options[mode]', 'eq', tool_uploadcourse_processor::MODE_CREATE_ALL);
-        $mform->addHelpButton('options[updatemode]', 'updatemode', 'tool_bulkenrollment');
-
-        $mform->addElement('selectyesno', 'options[allowdeletes]', get_string('allowdeletes', 'tool_bulkenrollment'));
-        $mform->setDefault('options[allowdeletes]', 0);
-        $mform->hideIf('options[allowdeletes]', 'options[mode]', 'eq', tool_uploadcourse_processor::MODE_CREATE_NEW);
-        $mform->hideIf('options[allowdeletes]', 'options[mode]', 'eq', tool_uploadcourse_processor::MODE_CREATE_ALL);
-        $mform->addHelpButton('options[allowdeletes]', 'allowdeletes', 'tool_bulkenrollment');
-
-        $mform->addElement('selectyesno', 'options[allowrenames]', get_string('allowrenames', 'tool_bulkenrollment'));
-        $mform->setDefault('options[allowrenames]', 0);
-        $mform->hideIf('options[allowrenames]', 'options[mode]', 'eq', tool_uploadcourse_processor::MODE_CREATE_NEW);
-        $mform->hideIf('options[allowrenames]', 'options[mode]', 'eq', tool_uploadcourse_processor::MODE_CREATE_ALL);
-        $mform->addHelpButton('options[allowrenames]', 'allowrenames', 'tool_bulkenrollment');
-
-        $mform->addElement('selectyesno', 'options[allowresets]', get_string('allowresets', 'tool_bulkenrollment'));
-        $mform->setDefault('options[allowresets]', 0);
-        $mform->hideIf('options[allowresets]', 'options[mode]', 'eq', tool_uploadcourse_processor::MODE_CREATE_NEW);
-        $mform->hideIf('options[allowresets]', 'options[mode]', 'eq', tool_uploadcourse_processor::MODE_CREATE_ALL);
-        $mform->addHelpButton('options[allowresets]', 'allowresets', 'tool_bulkenrollment');
+        
     }
 
 }
