+++ admin/tool/bulkenrollment/cli/bulkenrollment.php
@@ -17,7 +17,7 @@
 /**
  * CLI Bulk course registration script from a comma separated file.
  *
- * @package    tool_uploadcourse
+ * @package    tool_bulkenrollment
  * @copyright  2012 Piers Harding
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
@@ -85,7 +85,7 @@ Options:
 
 
 Example:
-\$sudo -u www-data /usr/bin/php admin/tool/uploadcourse/cli/uploadcourse.php --mode=createnew \\
+\$sudo -u www-data /usr/bin/php admin/tool/bulkenrollment/cli/bulkenrollment.php --mode=createnew \\
        --updatemode=dataonly --file=./courses.csv --delimiter=comma
 ";
 
@@ -105,10 +105,10 @@ $processoroptions = array(
 
 // Confirm that the mode is valid.
 $modes = array(
-    'createnew' => tool_uploadcourse_processor::MODE_CREATE_NEW,
-    'createall' => tool_uploadcourse_processor::MODE_CREATE_ALL,
-    'createorupdate' => tool_uploadcourse_processor::MODE_CREATE_OR_UPDATE,
-    'update' => tool_uploadcourse_processor::MODE_UPDATE_ONLY
+    'createnew' => tool_bulkenrollment_processor::MODE_CREATE_NEW,
+    'createall' => tool_bulkenrollment_processor::MODE_CREATE_ALL,
+    'createorupdate' => tool_bulkenrollment_processor::MODE_CREATE_OR_UPDATE,
+    'update' => tool_bulkenrollment_processor::MODE_UPDATE_ONLY
 );
 if (!isset($options['mode']) || !isset($modes[$options['mode']])) {
     echo get_string('invalidmode', 'tool_bulkenrollment')."\n";
@@ -119,13 +119,13 @@ $processoroptions['mode'] = $modes[$options['mode']];
 
 // Check that the update mode is valid.
 $updatemodes = array(
-    'nothing' => tool_uploadcourse_processor::UPDATE_NOTHING,
-    'dataonly' => tool_uploadcourse_processor::UPDATE_ALL_WITH_DATA_ONLY,
-    'dataordefaults' => tool_uploadcourse_processor::UPDATE_ALL_WITH_DATA_OR_DEFAUTLS,
-    'missingonly' => tool_uploadcourse_processor::UPDATE_MISSING_WITH_DATA_OR_DEFAUTLS
+    'nothing' => tool_bulkenrollment_processor::UPDATE_NOTHING,
+    'dataonly' => tool_bulkenrollment_processor::UPDATE_ALL_WITH_DATA_ONLY,
+    'dataordefaults' => tool_bulkenrollment_processor::UPDATE_ALL_WITH_DATA_OR_DEFAUTLS,
+    'missingonly' => tool_bulkenrollment_processor::UPDATE_MISSING_WITH_DATA_OR_DEFAUTLS
 );
-if (($processoroptions['mode'] === tool_uploadcourse_processor::MODE_CREATE_OR_UPDATE ||
-        $processoroptions['mode'] === tool_uploadcourse_processor::MODE_UPDATE_ONLY)
+if (($processoroptions['mode'] === tool_bulkenrollment_processor::MODE_CREATE_OR_UPDATE ||
+        $processoroptions['mode'] === tool_bulkenrollment_processor::MODE_UPDATE_ONLY)
         && (!isset($options['updatemode']) || !isset($updatemodes[$options['updatemode']]))) {
     echo get_string('invalideupdatemode', 'tool_bulkenrollment')."\n";
     echo $help;
@@ -188,8 +188,8 @@ cron_setup_user();
 
 // Let's get started!
 $content = file_get_contents($options['file']);
-$importid = csv_import_reader::get_new_iid('uploadcourse');
-$cir = new csv_import_reader($importid, 'uploadcourse');
+$importid = csv_import_reader::get_new_iid('bulkenrollment');
+$cir = new csv_import_reader($importid, 'bulkenrollment');
 $readcount = $cir->load_csv_content($content, $options['encoding'], $options['delimiter']);
 unset($content);
 if ($readcount === false) {
@@ -197,5 +197,5 @@ if ($readcount === false) {
 } else if ($readcount == 0) {
     throw new \moodle_exception('csvemptyfile', 'error', '', $cir->get_error());
 }
-$processor = new tool_uploadcourse_processor($cir, $processoroptions, $defaults);
-$processor->execute(new tool_uploadcourse_tracker(tool_uploadcourse_tracker::OUTPUT_PLAIN));
+$processor = new tool_bulkenrollment_processor($cir, $processoroptions, $defaults);
+$processor->execute(new tool_bulkenrollment_tracker(tool_bulkenrollment_tracker::OUTPUT_PLAIN));
