+++ admin/tool/bulkenrollment/index.php
@@ -17,7 +17,7 @@
 /**
  * Bulk course registration script from a comma separated file.
  *
- * @package    tool_uploadcourse
+ * @package    tool_bulkenrollment
  * @copyright  2011 Piers Harding
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
@@ -34,27 +34,27 @@ $previewrows = optional_param('previewrows', 10, PARAM_INT);
 $returnurl = new moodle_url('/admin/tool/bulkenrollment/index.php');
 
 if (empty($importid)) {
-    $mform1 = new tool_uploadcourse_step1_form();
+    $mform1 = new tool_bulkenrollment_step1_form();
     if ($form1data = $mform1->get_data()) {
-        $importid = csv_import_reader::get_new_iid('uploadcourse');
-        $cir = new csv_import_reader($importid, 'uploadcourse');
+        $importid = csv_import_reader::get_new_iid('bulkenrollment');
+        $cir = new csv_import_reader($importid, 'bulkenrollment');
         $content = $mform1->get_file_content('coursefile');
         $readcount = $cir->load_csv_content($content, $form1data->encoding, $form1data->delimiter_name);
         unset($content);
         if ($readcount === false) {
-            throw new \moodle_exception('csvfileerror', 'tool_uploadcourse', $returnurl, $cir->get_error());
+            throw new \moodle_exception('csvfileerror', 'tool_bulkenrollment', $returnurl, $cir->get_error());
         } else if ($readcount == 0) {
             throw new \moodle_exception('csvemptyfile', 'error', $returnurl, $cir->get_error());
         }
     } else {
         echo $OUTPUT->header();
-        echo $OUTPUT->heading_with_help(get_string('uploadcourses', 'tool_uploadcourse'), 'uploadcourses', 'tool_uploadcourse');
+        echo $OUTPUT->heading_with_help(get_string('bulkenrollments', 'tool_bulkenrollment'), 'bulkenrollments', 'tool_bulkenrollment');
         $mform1->display();
         echo $OUTPUT->footer();
         die();
     }
 } else {
-    $cir = new csv_import_reader($importid, 'uploadcourse');
+    $cir = new csv_import_reader($importid, 'bulkenrollment');
 }
 
 // Data to set in the form.
@@ -66,7 +66,7 @@ if (!empty($form1data)) {
     }
 }
 $context = context_system::instance();
-$mform2 = new tool_uploadcourse_step2_form(null, array('contextid' => $context->id, 'columns' => $cir->get_columns(),
+$mform2 = new tool_bulkenrollment_step2_form(null, array('contextid' => $context->id, 'columns' => $cir->get_columns(),
     'data' => $data));
 
 // If a file has been uploaded, then process it.
@@ -76,36 +76,19 @@ if ($form2data = $mform2->is_cancelled()) {
 } else if ($form2data = $mform2->get_data()) {
 
     $options = (array) $form2data->options;
-    $defaults = (array) $form2data->defaults;
 
-    // Custom field defaults.
-    $customfields = tool_uploadcourse_helper::get_custom_course_field_names();
-    foreach ($customfields as $customfield) {
-        $defaults[$customfield] = $form2data->{$customfield};
-    }
 
-    // Restorefile deserves its own logic because formslib does not really appreciate
-    // when the name of a filepicker is an array...
-    $options['restorefile'] = '';
-    if (!empty($form2data->restorefile)) {
-        $options['restorefile'] = $mform2->save_temp_file('restorefile');
-    }
-    $processor = new tool_uploadcourse_processor($cir, $options, $defaults);
+    $processor = new tool_bulkenrollment_processor($cir, $options);
 
     echo $OUTPUT->header();
     if (isset($form2data->showpreview)) {
-        echo $OUTPUT->heading(get_string('uploadcoursespreview', 'tool_uploadcourse'));
-        $processor->preview($previewrows, new tool_uploadcourse_tracker(tool_uploadcourse_tracker::OUTPUT_HTML));
+        echo $OUTPUT->heading(get_string('bulkenrollmentspreview', 'tool_bulkenrollment'));
+        $processor->preview($previewrows, new tool_bulkenrollment_tracker(tool_bulkenrollment_tracker::OUTPUT_HTML));
         $mform2->display();
     } else {
-        echo $OUTPUT->heading(get_string('uploadcoursesresult', 'tool_uploadcourse'));
-        $processor->execute(new tool_uploadcourse_tracker(tool_uploadcourse_tracker::OUTPUT_HTML));
+        echo $OUTPUT->heading(get_string('bulkenrollmentsresult', 'tool_bulkenrollment'));
+        $processor->execute(new tool_bulkenrollment_tracker(tool_bulkenrollment_tracker::OUTPUT_HTML));
         echo $OUTPUT->continue_button($returnurl);
-    }
-
-    // Deleting the file after processing or preview.
-    if (!empty($options['restorefile'])) {
-        @unlink($options['restorefile']);
     }
 
 } else {
@@ -115,12 +98,12 @@ if ($form2data = $mform2->is_cancelled()) {
         $options = (array)$submitteddata->options;
     } else {
         // Weird but we still need to provide a value, setting the default step1_form one.
-        $options = array('mode' => tool_uploadcourse_processor::MODE_CREATE_NEW);
+        $options = array('mode' => tool_bulkenrollment_tracker::MODE_CREATE_NEW);
     }
-    $processor = new tool_uploadcourse_processor($cir, $options, array());
+    $processor = new tool_bulkenrollment_tracker($cir, $options);
     echo $OUTPUT->header();
-    echo $OUTPUT->heading(get_string('uploadcoursespreview', 'tool_uploadcourse'));
-    $processor->preview($previewrows, new tool_uploadcourse_tracker(tool_uploadcourse_tracker::OUTPUT_HTML));
+    echo $OUTPUT->heading(get_string('bulkenrollmentspreview', 'tool_bulkenrollment'));
+    $processor->preview($previewrows, new tool_bulkenrollment_tracker(tool_bulkenrollment_tracker::OUTPUT_HTML));
     $mform2->display();
 }
