+++ admin/tool/bulkenrollment/version.php
@@ -17,7 +17,7 @@
 /**
  * Plugin version info.
  *
- * @package    tool_uploadcourse
+ * @package    tool_bulkenrollment
  * @copyright  2011 Piers Harding
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
